Fix range around wlist
Add new method addNotFailedBuild which is checking if build in range around whitelist is crashed or not, if the build is crashed then is replaced with more remote one. This is done by two for cycles. They are expending their range (this depens on return value of new method) Method is also adding not crashed builds to the list.